Do not communicate with server when creating WebDavCollection objects
PROPFIND will now first be performed when the properties are queried. It
is burdensome for an application to delay creation of WebDavCollection /
AddressbookCollection objects in case no communication to the server
will be needed, so we move the delayed querying of properties to the
lib.

Also moved supportsSyncCollection to WebDavCollection, as the
sync-collection report is not CardDAV specific but a WebDAV extension.