Fix: try provided path on  addressbook discovery
The discovery service will first try the standard discovery mechanisms,
assuming that the service administrator knows best where the services
are located. However, if discovery mechanisms have not been properly set
up admin side, it should be possible to provide a full uri as input for
that the discovery works. The path component of the input URI was
ignored so far. Now it is tried if everything else failed.