Fix: no exceptions in discovery methods of Account
According to their interface, the functions to query principal URI,
addressbook home URI and the addressbooks' URIs should indicate errors
by return value, as these are to be expected during discovery attempts.
This was not the case and exceptions where thrown, causing the Discovery
service to abort. The implementation has now been fixed to behave as
documented in the interface.