SyncHandler may get passed NULL changed cards
In case a card is reported as changed by the server, but an error occurs
when retrieving or parsing that VCard, the addressObjectChanged()
function is nevertheless invoked with valid $uri and $etag parameters,
but the $card will be null in that case. This enables clients to be
informed about the problem and deal with it as they see fit.

This is now reflected in the interface, where previously $card had a
non-null type documented.