Use authentication only within domain
When a CardDavClient (and therefore HTTP Client) is created, it is given
authentication info and a base URI. The client may be used to send
requests to any URL, including those with different domain. It may not
be desired in such cases that the authentication information is
transmitted to a different domain. (We assume that the domain is under
control of a trusted party that the user intended to auth with).

This only affects the URI given by the using application. In particular,
if redirections occur during a request, authentication will be used if
it was used for the original URI (the redirection is provided by a
trusted domain).

The particular background: VCards may include URIs, for example for the
photo. These may be somewhat easily injected into a user's addressbook
(i.e. send a VCard and import it). In this case, it might be dangerous
to send the credentials to a server hosting such an URI. If the client
is willing to do this, they have to create a separate client object with
a suited base URI.