Make CardDavClient URI comparison more tolerant
Particularly, disregard the initial slash as well in the comparison.
iCloud was found to return URIs in a 507 status element that did not
include the initial slash.

Example: /xxx/yyy/ compares equals to xxx/yyy

It should not cause harm with other servers, as servers always seem to
return URIs relative to the server root.