Support servers with multiple addressbook home locations for one principal
This changes a public function of the Account class, however, it was not marked as part of the public API and thus the change is not considered a backwards incompatible API break. The Discovery service is adapted to support multiple addressbook homes, but the interface of the service is not changed.